Add in initial classes for solution:
CoOrds : simple data type to represent co-ordinates within a grid
Map : is an interface that allows users to create a new map with "template"
methods to interact with that map.
SimpleGrid : implements a map. In this case it will be a simple X x Y grid.
(either square or rectangle)

Currently only default (generated) unit tests shown.